feat(auth): Enhance two-factor authentication flow and add MFA method selection
