+++ public/locale/en.json
@@ -2583,7 +2583,9 @@
   "uploading": "Uploading",
   "use": "Use",
   "use_address_as_permanent": "Use this address for permanent address",
+  "use_auth_app": "Use authenticator app",
   "use_authenticator_app": "Use an authenticator app or browser extension to scan.",
+  "use_backup_code": "Use a backup code",
   "use_phone_number_for_emergency": "Use this phone number for emergency contact",
   "user_add_error": "Error while adding User",
   "user_added_successfully": "User added successfully",
