+++ src/components/Auth/Authenticate.tsx
@@ -1,6 +1,5 @@
-import careConfig from "@careConfig";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Link, navigate } from "raviger";
+import { navigate } from "raviger";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
@@ -21,17 +20,28 @@ import { Input } from "@/components/ui/input";
 
 import { useAuthContext } from "@/hooks/useAuthUser";
 
+import { MFAMethod, MFAOption } from "@/types/auth/otp";
+
+import { AuthHero } from "./AuthHero";
+
 export const Authenticate = () => {
-  const { urls, stateLogo, customLogo, customLogoAlt } = careConfig;
-  const customDescriptionHtml = __CUSTOM_DESCRIPTION_HTML__;
-  const logos = [stateLogo, customLogo].filter(
-    (logo) => logo?.light || logo?.dark,
-  );
   const { t } = useTranslation();
   const [error, setError] = useState<string>("");
-  const method = localStorage.getItem("mfa_method") || "totp";
-  const [recoveryModal, setRecoveryModal] = useState(false);
+  const [currentMethod, setCurrentMethod] = useState<MFAMethod>("totp");
   const { verifyMFA, isVerifyingMFA } = useAuthContext();
+
+  // Available MFA methods configuration
+  const mfaOptions: MFAOption[] = [
+    {
+      id: "totp",
+      label: t("use_auth_app"),
+    },
+    {
+      id: "backup",
+      label: t("use_backup_code"),
+    },
+    // Add more methods here as we make them available
+  ];
 
   // Form validation schema
   const formSchema = z.object({
@@ -68,7 +78,7 @@ export const Authenticate = () => {
 
     try {
       await verifyMFA({
-        method: recoveryModal ? "backup" : method,
+        method: currentMethod,
         code: values.code,
         temp_token,
       });
@@ -80,123 +90,21 @@ export const Authenticate = () => {
     }
   });
 
-  const accesWays: readonly string[] = ["Use a recovery code"];
-  const recoveryWays: readonly string[] = ["Use authenticator app"];
+  // Get available alternative methods based on current method
+  const alternativeMethods = mfaOptions.filter(
+    (option) => option.id !== currentMethod,
+  );
 
-  function handleRedirect(way: string): void {
-    if (way === "Use a recovery code") {
-      setRecoveryModal(true);
-    } else if (way === "Use authenticator app") {
-      setRecoveryModal(false);
-    }
-  }
+  // Handle method change
+  const handleMethodChange = (method: MFAMethod) => {
+    setCurrentMethod(method);
+    form.reset();
+    setError("");
+  };
 
   return (
     <div className="relative flex min-h-screen flex-col md:h-screen md:flex-row">
-      {/* Hero Section */}
-      <div className="login-hero order-last relative flex flex-auto flex-col justify-between p-6 md:order-first md:h-full md:w-[calc(50%+130px)] md:flex-none md:p-0 md:px-16 md:pr-[calc(4rem+130px)]">
-        <div></div>
-        <div className="mt-4 flex flex-col items-start rounded-lg py-4 md:mt-12">
-          <div className="mb-4 hidden items-center gap-6 md:flex">
-            {logos.map((logo, index) =>
-              logo && logo.light ? (
-                <div key={index} className="flex items-center">
-                  <img
-                    src={logo.light}
-                    className="h-16 rounded-lg py-3"
-                    alt="state logo"
-                  />
-                </div>
-              ) : null,
-            )}
-            {logos.length === 0 && (
-              <a
-                href={urls.ohcn}
-                className="inline-block"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <img
-                  src={customLogoAlt?.light ?? "/images/ohc_logo_light.svg"}
-                  className="h-8"
-                  alt="Open Healthcare Network logo"
-                />
-              </a>
-            )}
-          </div>
-          <div className="max-w-lg">
-            <h1 className="text-4xl font-black leading-tight tracking-wider text-white lg:text-5xl">
-              {t("care")}
-            </h1>
-            {customDescriptionHtml ? (
-              <div className="py-6">
-                <div
-                  className="max-w-xl text-secondary-400"
-                  dangerouslySetInnerHTML={{
-                    __html: __CUSTOM_DESCRIPTION_HTML__,
-                  }}
-                />
-              </div>
-            ) : (
-              <div className="max-w-xl py-6 pl-1 text-base font-semibold text-secondary-400 md:text-lg lg:text-xl">
-                {t("goal")}
-              </div>
-            )}
-          </div>
-        </div>
-        <div className="mb-6 flex items-center">
-          <div className="max-w-lg text-xs md:text-sm">
-            <div className="mb-2 ml-1 flex items-center gap-4">
-              <a
-                href="https://www.digitalpublicgoods.net/r/care"
-                rel="noopener noreferrer"
-                target="_blank"
-              >
-                <img
-                  src="https://cdn.ohc.network/dpg-logo.svg"
-                  className="h-12"
-                  alt="Logo of Digital Public Goods Alliance"
-                />
-              </a>
-              <div className="ml-2 h-8 w-px rounded-full bg-white/50" />
-              <a href={urls.ohcn} rel="noopener noreferrer" target="_blank">
-                <img
-                  src="/images/ohc_logo_light.svg"
-                  className="inline-block h-10"
-                  alt="Open Healthcare Network logo"
-                />
-              </a>
-            </div>
-            <a
-              href={urls.ohcn}
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-secondary-500"
-            >
-              {t("footer_body")}
-            </a>
-            <div className="mx-auto mt-2">
-              <a
-                href={urls.github}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-primary-400 hover:text-primary-500"
-              >
-                {t("contribute_github")}
-              </a>
-              <span className="mx-2 text-primary-400">|</span>
-              <Link
-                href="/licenses"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="text-primary-400 hover:text-primary-500"
-              >
-                {t("third_party_software_licenses")}
-              </Link>
-            </div>
-          </div>
-        </div>
-      </div>
+      <AuthHero />
       {/* Login Forms Section */}
       <div className="login-hero-form py-16 w-full md:mt-0 md:h-full md:w-1/2">
         <div className="relative h-full items-center flex justify-center md:flex">
@@ -218,16 +126,18 @@ export const Authenticate = () => {
                           <FormControl>
                             <Input
                               placeholder={
-                                recoveryModal ? "XXXXXXXX" : "XXXXXX"
+                                currentMethod === "backup"
+                                  ? "XXXXXXXX"
+                                  : "XXXXXX"
                               }
                               {...field}
-                              maxLength={recoveryModal ? 8 : 6}
+                              maxLength={currentMethod === "backup" ? 8 : 6}
                               autoComplete="one-time-code"
                               className="tracking-[0.1em] placeholder:text-gray-500/50"
                             />
                           </FormControl>
                           <FormLabel className="mt-3">
-                            {recoveryModal
+                            {currentMethod === "backup"
                               ? t("enter_recovery_code")
                               : t("enter_2fa_code")}
                           </FormLabel>
@@ -240,7 +150,7 @@ export const Authenticate = () => {
                       variant="primary"
                       disabled={
                         isVerifyingMFA ||
-                        codeValue.length < (recoveryModal ? 8 : 6)
+                        codeValue.length < (currentMethod === "backup" ? 8 : 6)
                       }
                     >
                       {isVerifyingMFA ? t("verifying") : t("verify")}{" "}
@@ -256,20 +166,20 @@ export const Authenticate = () => {
 
                     <div className="mt-5 text-center">
                       <p className="text-sm text-gray-500 font-base">
-                        {recoveryModal ? "" : t("cant_access_code")}
+                        {currentMethod === "backup"
+                          ? ""
+                          : t("cant_access_code")}
                       </p>
                       <ul className="list-disc inline-flex justify-center w-full">
-                        {(recoveryModal ? recoveryWays : accesWays).map(
-                          (way: string, idx: number) => (
-                            <li
-                              key={idx}
-                              className="text-sm font-medium text-primary-500 hover:underline cursor-pointer"
-                              onClick={() => handleRedirect(way)}
-                            >
-                              {way}
-                            </li>
-                          ),
-                        )}
+                        {alternativeMethods.map((method) => (
+                          <li
+                            key={method.id}
+                            onClick={() => handleMethodChange(method.id)}
+                            className="text-sm font-medium text-primary-500 hover:underline cursor-pointer"
+                          >
+                            {method.label}
+                          </li>
+                        ))}
                       </ul>
                     </div>
                   </form>
